+++ src/tools/ModPrime.java
@@ -8,7 +8,7 @@ public class ModPrime {
 	 * @param b
 	 * @return
 	 */
-	public static int calc(int a, int b) {
-	    return a - b * ((int) Math.floor(a / (float) b + 1 / 2.0));
+	public static long calc(long a, long b) {
+	    return a - b * ((long) Math.floor(a / (float) b + 1 / 2.0));
 	}
 }
